fix: Form string correctly
Fixes the incorrectly formed string for the case when there are no additional stacks chosen. In the current implementation the string 'railyard hash -b <self.base> ' is formed with the trailing space. When applying `split` it returns list ['railyard', 'hash', '-b', '<self.base>', '']. The last element of the list is passed to subprocess as the empty argument, and `railyard` CLI fails with `Error: Unexpected argument:`. As a result `tag` variable is not set in line 42 and spawn fails.